+++ changelog.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.7] - 2020.07.12
+
+**Added**
+- Svgs: add, criss-cross.
+- Components: NewSetBtn.
+
+**Deleted**
+- new.svg.
+
+## [0.0.6] - 2020.07.10
+
+**Changes**
+- Swap from css styling to styled-components
+
+**Deleted**
+- All style files *.scss
+
 ## [0.0.5] - 2020.06.12
 
 **Added**